rename: qna 폴더 domain 안으로 이동
