+++ src/main/java/com/example/tomyongji/domain/qna/controller/QnaAnswerController.java
@@ -1,10 +1,10 @@
-package com.example.tomyongji.qna.controller;
+package com.example.tomyongji.domain.qna.controller;
 
 import com.example.tomyongji.global.common.response.ApiResponse;
-import com.example.tomyongji.qna.dto.request.AnswerSaveDto;
-import com.example.tomyongji.qna.dto.response.AnswerDto;
-import com.example.tomyongji.qna.dto.response.PageResponseDto;
-import com.example.tomyongji.qna.service.QnaAnswerService;
+import com.example.tomyongji.domain.qna.dto.request.AnswerSaveDto;
+import com.example.tomyongji.domain.qna.dto.response.AnswerDto;
+import com.example.tomyongji.domain.qna.dto.response.PageResponseDto;
+import com.example.tomyongji.domain.qna.service.QnaAnswerService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
