+++ src/main/java/com/example/tomyongji/domain/qna/entity/QnaAnswer.java
@@ -1,0 +1,43 @@
+package com.example.tomyongji.domain.qna.entity;
+
+import com.example.tomyongji.domain.receipt.entity.StudentClub;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter @Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class QnaAnswer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "student_club_id")
+    private StudentClub writer;
+    private String content;
+
+    @Column(updatable = false)
+    private LocalDateTime createdTime;
+
+    private LocalDateTime updatedTime;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private QnaQuestion question;
+
+    @PrePersist
+    public void prePersist() {
+        this.createdTime = LocalDateTime.now();
+        this.updatedTime = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    public void preUpdate() {
+        this.updatedTime = LocalDateTime.now();
+    }
+}
