+++ src/main/java/com/example/tomyongji/domain/qna/repository/QnaAnswerRepository.java
@@ -1,12 +1,10 @@
-package com.example.tomyongji.qna.repository;
+package com.example.tomyongji.domain.qna.repository;
 
-import com.example.tomyongji.qna.entity.QnaAnswer;
+import com.example.tomyongji.domain.qna.entity.QnaAnswer;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface QnaAnswerRepository extends JpaRepository<QnaAnswer,Long> {
